added constructors in Gmap.h for structs vertex, edge and face

+++ hw/01/cpp/src/main.cpp
@@ -40,7 +40,7 @@ int main(int argc, const char * argv[]) {
 
         if (word == "f") {
           std::vector<int> face;
-          while (iss >> word) face.push_back(std::stof(word));
+          while (iss >> word) face.push_back(std::stoi(word));
           face_indices.push_back(face);
         }
